pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "adverse-weather-semantic-segmentation-robustness-benchmark"
version = "1.0.0"
description = "Benchmarks and improves semantic segmentation model robustness under synthetically degraded weather conditions"
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    {name = "Alireza Shojaei", email = "alireza.shojaei@example.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Image Processing",
]
requires-python = ">=3.8"
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "timm>=0.9.0",
    "albumentations>=1.3.0",
    "opencv-python>=4.8.0",
    "numpy>=1.24.0",
    "scipy>=1.11.0",
    "scikit-learn>=1.3.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "Pillow>=10.0.0",
    "tqdm>=4.65.0",
    "pyyaml>=6.0",
    "tensorboard>=2.13.0",
    "mlflow>=2.5.0",
    "pandas>=2.0.0",
    "transformers>=4.30.0",
    "segmentation-models-pytorch>=0.3.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
]
wandb = ["wandb>=0.15.0"]
plotly = ["plotly>=5.15.0"]

[project.urls]
Homepage = "https://github.com/alireza-shojaei/adverse-weather-semantic-segmentation-robustness-benchmark"
Repository = "https://github.com/alireza-shojaei/adverse-weather-semantic-segmentation-robustness-benchmark"
Documentation = "https://github.com/alireza-shojaei/adverse-weather-semantic-segmentation-robustness-benchmark#readme"

[project.scripts]
train-adverse-weather = "adverse_weather_semantic_segmentation_robustness_benchmark.scripts.train:main"
evaluate-adverse-weather = "adverse_weather_semantic_segmentation_robustness_benchmark.scripts.evaluate:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"adverse_weather_semantic_segmentation_robustness_benchmark" = ["configs/*.yaml"]

[tool.black]
line-length = 100
target-version = ["py38", "py39", "py310", "py311"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | \.env
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=src/adverse_weather_semantic_segmentation_robustness_benchmark --cov-report=html --cov-report=term-missing"
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.coverage.run]
source = ["src/adverse_weather_semantic_segmentation_robustness_benchmark"]
omit = [
    "*/tests/*",
    "*/test_*.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]